requirements.txt
humanize
uvloop
discord.py
wavelink
git+https://github.com/Rapptz/discord-ext-menus.git
pillow
requests